consolidate CompressionUtils.java since now in the same jar
